+++ tests/utilities/test_utilities_base.py
@@ -9,7 +9,6 @@ import pandas as pd
 
 from gaps import TEST_DATA_DIR
 from gaps.utilities import (
-    CaseInsensitiveEnum,
     recursively_update_dict,
     resolve_path,
     project_points_from_container_or_slice,
@@ -19,55 +18,6 @@ from gaps.exceptions import gapsValueError
 
 TEST_1_ATTRS_1 = {"job_name": "test1", "job_status": "R", "run_id": 1234}
 TEST_1_ATTRS_2 = {"job_name": "test1", "job_status": "successful"}
-
-
-def test_case_insensitive_enum():
-    """Tets subclassing the case insensitive enum"""
-
-    # pylint: disable=too-few-public-methods
-    class TestEnum(CaseInsensitiveEnum):
-        """A test enum."""
-
-        HELLO = "hello"
-        THERE = "THERE"
-        THIS = "ThIS"
-
-        @classmethod
-        def _new_post_hook(cls, obj, value):
-            obj.my_test_len = len(value)
-            return obj
-
-    assert f"{TestEnum.HELLO}" == "hello"
-    assert f"{TestEnum.THERE}" == "there"
-    assert f"{TestEnum.THIS}" == "this"
-
-    for text in ["hello", " HELLO", " HeLlo  "]:
-        assert TestEnum(text) == TestEnum.HELLO
-
-    for text in ["there", " THERE", " ThEre  "]:
-        assert TestEnum(text) == TestEnum.THERE
-
-    for text in ["this", " THIS", " ThIs  "]:
-        assert TestEnum(text) == TestEnum.THIS
-
-    with pytest.raises(ValueError):
-        TestEnum("dne")
-
-    with pytest.raises(ValueError):
-        TestEnum("DNE")
-
-    with pytest.raises(ValueError):
-        TestEnum("DnE")
-
-    with pytest.raises(ValueError):
-        TestEnum(None)
-
-    # pylint: disable=no-member
-    assert TestEnum.HELLO.my_test_len == 5
-    assert TestEnum.THERE.my_test_len == 5
-    assert TestEnum.THIS.my_test_len == 4
-
-    assert TestEnum.members_as_str() == {"hello", "there", "this"}
 
 
 def test_project_points_from_container_or_slice():
